Simula dados do modelo beta ordinal com dispersão variável

Esta função simula dados de uma variável beta ordinal, aplicando diferentes funções de ligação tanto para betas como para zetas de phi

Usage

betaregesc_simula_dados_z(
  formula_x = ~x1 + x2,
  formula_z = ~z1 + z2,
  dados,
  betas = c(0, 0.5, -0.2),
  zetas = c(1, 0.5, 0.2),
  link = "logit",
  link_phi = "log",
  ncuts = 100,
  type = "m",
  lim = 0.5
)

Arguments

formula_x

Fórmula para expressar a relação das preditoras X1, X2, Xn relacionadas com os betas. Ela deve ser referenciada em Y. Ex. formula = ~X1 + X2. Isso porque a variável resposta é intervalar. Veja os detalhes para mais informação.

formula_z

Fórmula para expressar a relação das preditoras Z1, Z2, Zn relacionadas com os phi's no modelo se regressão beta com preditoas em phi. Ela deve ser referenciada em Y. Ex. formula = ~Z1 + Z2. Isso porque a variável resposta é intervalar. Veja os detalhes para mais informação.

dados

Um conjunto de dados que contém a variável dependente e as variáveis independentes especificadas nas fórmulas.

betas

Vetor de betas associados aos preditores de X utilizados para simular da beta.

zetas

Vetor de zetas associados aos preditores de Z utilizados para simular da beta com dispersão variável.

link

Nome da função de ligação a ser usada para as preditoras X1, X2, ..., Xn. Pode ser uma das seguintes: "logit", "probit", "cauchit", "cloglog" ou "identity". O padrão é "logit".

link_phi

Nome da função de ligação a ser usada para as preditoras Z1, Z2, ..., Zn relacionadas com phi. Pode ser uma das seguintes: "log", "sqrt e "identity". O padrão é "log".

ncuts

Número de cortes para a variável ordinal. O padrão é 100.

type

Tipo de intervalo. "m" = meio; "l" = esquerda e "r" = direita.

lim

Limite numérico a ser utilizado para ajustar os intervalos. O padrão é 0.5.

Value

Retorna um data.frame contendo os dados simulados da variável beta ordinal e as variáveis independentes.

Examples

# Criar um conjunto de dados de exemplo
set.seed(421)
n <- 50
dados <- data.frame(
x1 = rnorm(n), x2 = rnorm(n),
z1 = runif(n), z2 = runif(n))
fx <- ~ x1 + x2
fz <- ~ z1
dados_simulados <- betaregesc_simula_dados_z(
formula_x = fx,
formula_z = fz,
dados = dados,
betas = c(0.2, -0.5, 0.3),
zetas = c(1, 1.2),
link = "logit",
link_phi = "log",
ncuts = 100,
type = "m")